Fix precise input handling on Hashlink.

+++ project/src/backend/sdl/SDLSystem.cpp
@@ -87,7 +87,7 @@ namespace lime {
 	}
 
 
-	long System::GetTicks () {
+	int System::GetTicks () {
 
 		return SDL_GetTicks ();
